+++ packages/fiori-elements-writer/test/lrop.test.ts
@@ -13,6 +13,15 @@ import {
 
 const TEST_NAME = 'lropTemplates';
 
+jest.mock('read-pkg-up', () => ({
+    sync: jest.fn().mockReturnValue({
+        packageJson: {
+            name: 'mocked-package-name',
+            version: '9.9.9-mocked'
+        }
+    })
+}));
+
 describe(`Fiori Elements template: ${TEST_NAME}`, () => {
     const curTestOutPath = join(testOutputDir, TEST_NAME);
 
@@ -101,6 +110,24 @@ describe(`Fiori Elements template: ${TEST_NAME}`, () => {
                         sourceTemplate: {
                             version: '1.2.3-test',
                             id: 'test-fe-template',
+                            toolsId: 'toolsId:1234abcd'
+                        }
+                    }
+                }),
+                service: v2Service
+            } as FioriElementsApp<LROPSettings>
+        },
+        {
+            name: 'lropV2_set_toolsId_only',
+            config: {
+                ...Object.assign(feBaseConfig('felrop5'), {
+                    template: {
+                        type: TemplateType.ListReportObjectPage,
+                        settings: v2TemplateSettings
+                    },
+                    app: {
+                        ...feBaseConfig('felrop4').app,
+                        sourceTemplate: {
                             toolsId: 'toolsId:1234abcd'
                         }
                     }