+++ packages/fiori-freestyle-writer/test/basic.test.ts
@@ -7,6 +7,15 @@ import { BasicAppSettings } from '../src/types';
 
 const TEST_NAME = 'basicTemplate';
 
+jest.mock('read-pkg-up', () => ({
+    sync: jest.fn().mockReturnValue({
+        packageJson: {
+            name: 'mocked-package-name',
+            version: '9.9.9-mocked'
+        }
+    })
+}));
+
 describe(`Fiori freestyle template: ${TEST_NAME}`, () => {
     const curTestOutPath = join(testOutputDir, TEST_NAME);
 
@@ -71,6 +80,19 @@ describe(`Fiori freestyle template: ${TEST_NAME}`, () => {
             config: {
                 ...commonConfig,
                 appOptions: { loadReuseLibs: false }
+            },
+            settings: {}
+        },
+        {
+            name: 'basic_with_toolsId',
+            config: {
+                ...commonConfig,
+                app: {
+                    ...commonConfig.app,
+                    sourceTemplate: {
+                        toolsId: 'testToolsId:abcd1234'
+                    }
+                }
             },
             settings: {}
         }